GCC: Do not mess with progname in lrealpath.c
.. On Windows, lrealpath in libiberty calls
GetFullPathName and then CharLowerBuff. This
causes differences in the value of the path
between the build system and GCC internally
which causes (e.g.) sed to fail in glibc so
fix the dirseps back to  '/' (matching those
of any build system that can be used for GCC)
and keep the case as it was.

Patch supplied for GCC versions 4.8.2 and 4.9.2